Modul Keuangan (SPP)/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
APScheduler==3.10.4
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
flasgger==0.9.7.1
